rss-feed/requirements.txt
feedparser>=6.0.11
PyYAML>=6.0.1
requests>=2.32.3
python-dateutil>=2.9.0.post0
scikit-learn>=1.5.2
